llm_query_app/requirements.txt
streamlit>=1.35.0
pymupdf>=1.23.0
python-docx>=1.0.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
